🚚 Renamed `DefaultLayout` component and used index file to export
