+++ src/components/integrations-content.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import _ from 'lodash'
 
-import Layout from './layouts/default'
+import { Layout } from './layouts'
 import Integration from "./integration"
 import { Spirit } from './spirit-styles'
 import NavBar from './layouts/partials/navbar'
