+++ src/components/layouts/index.js
@@ -1,0 +1,2 @@
+export { default as Layout } from './DefaultLayout'
+// export { default as NavBar } from './NavBar'
