+++ src/pages/faq.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { graphql } from 'gatsby'
 
-import Layout from '../components/layouts/default'
+import { Layout } from '../components/layouts'
 import Link from 'gatsby-link'
 import { Spirit } from '../components/spirit-styles'
 import { MetaData } from '../components/meta'
