+++ src/pages/setup.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { graphql } from 'gatsby'
 
-import Layout from '../components/layouts/default'
+import { Layout } from '../components/layouts'
 import Icon from '../components/global/icon'
 import { Spirit } from '../components/spirit-styles'
 import Box from '../components/Box'
